Update to copier-cpp template version 0.2.4

+++ tests/bufferbus.test.cpp
@@ -5,15 +5,16 @@
 //         (See accompanying file LICENSE or copy at
 //           https://www.boost.org/LICENSE_1_0.txt)
 
-#include <dplx/dp/decoder/api.hpp>
-#include <dplx/dp/decoder/core.hpp>
-#include <dplx/dp/encoder/api.hpp>
-#include <dplx/dp/encoder/core.hpp>
+#include <dplx/dp.hpp>
+#include <dplx/dp/api.hpp>
+#include <dplx/dp/codecs/core.hpp>
 
 #include <dplx/dlog/log_bus.hpp>
 
 #include "boost-test.hpp"
 #include "test-utils.hpp"
+
+// NOLINTBEGIN(cppcoreguidelines-avoid-magic-numbers)
 
 namespace dlog_tests
 {
@@ -28,11 +29,11 @@ BOOST_AUTO_TEST_CASE(tmp)
 
     auto bufferbus = std::move(createRx).assume_value();
 
-    auto msgId = 0u;
+    auto msgId = 0U;
     dlog::bufferbus_handle::logger_token token{};
     for (;;)
     {
-        auto size = dp::encoded_size_of(msgId);
+        auto size = static_cast<unsigned>(dp::encoded_size_of(msgId));
         if (auto writeRx = bufferbus.write(token, size); writeRx.has_value())
         {
             auto encodeRx = dp::encode(writeRx.assume_value(), msgId);
@@ -75,3 +76,5 @@ BOOST_AUTO_TEST_CASE(tmp)
 BOOST_AUTO_TEST_SUITE_END()
 
 } // namespace dlog_tests
+
+// NOLINTEND(cppcoreguidelines-avoid-magic-numbers)
